+++ sources/kernel/kernel.h
@@ -47,7 +47,7 @@ uint32_t kernel_version(void);
 void     kernel_idle(void);
 uint32_t kernel_idle_time(void);
 void     kernel_tick(uint32_t time);
-uint32_t kernel_tick_count(void);
+uint64_t kernel_tick_count(void);
 
 /******************************************************************************
 * heap
